LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.drawing.style.invalid.spacing-expected.txt

2d.text.drawing.style.invalid.spacing
Testing letter spacing and word spacing with invalid units
Actual output:

FAIL Testing letter spacing and word spacing with invalid units assert_equals: ctx.letterSpacing === '0px' (got [undefined], expected 0px[string]) expected (string) "0px" but got (undefined) undefined
